feat: refactoring banner
