+++ package.json
@@ -27,6 +27,7 @@
     "eslint-config-next": "13.4.19",
     "eslint-config-prettier": "^9.0.0",
     "lodash": "^4.17.21",
+    "net": "^1.0.2",
     "next": "^13.5.4",
     "next-auth": "^4.23.1",
     "pretendard": "^1.3.9",
@@ -38,6 +39,7 @@
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
     "swr": "^2.2.4",
+    "tls": "^0.0.1",
     "web-push": "^3.6.6",
     "zustand": "^4.4.1"
   },
