+++ src/app/panel/MainPage.tsx
@@ -27,10 +27,11 @@ import useAuthStore from '@/states/useAuthStore'
 import useAxiosWithAuth from '@/api/config'
 import { AxiosInstance } from 'axios'
 import { IPagination } from '@/types/IPagination'
-import useMedia from '@/hook/useMedia'
 import webpush from 'web-push'
+import PwaInstallBanner from './PwaInstallBanner'
+import PushAlertBanner from './PushAlertBanner'
 
-interface BeforeInstallPromptEvent extends Event {
+export interface BeforeInstallPromptEvent extends Event {
   readonly platforms: string[]
   readonly userChoice: Promise<{
     outcome: 'accepted' | 'dismissed'
@@ -52,9 +53,6 @@ export interface IDetailOption {
 }
 
 const MainPage = ({ initData }: { initData: IPagination<IPost[]> }) => {
-  const { isPc } = useMedia()
-  const [isShowInstall, setIsShowInstall] = useState<boolean>(true)
-  const [isShowPush, setIsShowPush] = useState<boolean>(true)
   const [content, setContent] = useState<IPost[]>(initData?.content)
   const [page, setPage] = useState<number>(1)
   const [type, setType] = useState<ProjectType | undefined>(undefined) //'STUDY'
@@ -103,161 +101,6 @@ const MainPage = ({ initData }: { initData: IPagination<IPost[]> }) => {
       : defaultGetFetcher,
   )
 
-  const handleInstall = () => {
-    if (deferredPrompt) {
-      deferredPrompt.prompt()
-      deferredPrompt.userChoice.then((choiceResult) => {
-        console.log(choiceResult.outcome)
-        if (choiceResult.outcome === 'accepted') {
-          console.log('User accepted the install prompt')
-          setIsShowInstall(false)
-          localStorage.setItem('isShowInstall', 'false')
-        } else {
-          console.log('User dismissed the install prompt')
-        }
-      })
-    }
-  }
-
-  useEffect(() => {
-    if (localStorage && localStorage.getItem('isShowInstall') === 'false') {
-      setIsShowInstall(false)
-    }
-
-    if (localStorage && localStorage.getItem('isShowPush') === 'false') {
-      setIsShowPush(false)
-    }
-
-    const handleBeforeInstallPrompt = (e: BeforeInstallPromptEvent) => {
-      e.preventDefault()
-      setDeferredPrompt(e)
-      setIsShowInstall(true)
-    }
-
-    window.addEventListener('beforeinstallprompt', handleBeforeInstallPrompt)
-    window.addEventListener('appinstalled', () => {
-      console.log('installed')
-      setIsShowInstall(false)
-    })
-
-    // 테스트 용
-    window.addEventListener('push', (event) => {
-      const options = {
-        body: 'This is a test message!',
-        icon: 'images/icon.png',
-        badge: 'images/badge.png',
-      }
-      navigator.serviceWorker.ready.then((swReg) => {
-        swReg.showNotification('Test Push', options)
-      })
-    })
-  }, [deferredPrompt, isShowInstall, isShowPush])
-
-  const urlBase64ToUint8Array = (base64String: string) => {
-    const padding = '='.repeat((4 - (base64String.length % 4)) % 4)
-    const base64 = (base64String + padding)
-      .replace(/-/g, '+')
-      .replace(/_/g, '/')
-    const rawData = window.atob(base64)
-    const outputArray = new Uint8Array(rawData.length)
-
-    for (let i = 0; i < rawData.length; ++i) {
-      outputArray[i] = rawData.charCodeAt(i)
-    }
-
-    return outputArray
-  }
-
-  const displayNotification = () => {
-    if ('serviceWorker' in navigator && 'PushManager' in window) {
-      navigator.serviceWorker.ready.then((swReg) => {
-        swReg.showNotification('Hello world!')
-      })
-    }
-  }
-
-  const createPushSubscription = (swReg: ServiceWorkerRegistration) => {
-    // 테스트용
-    swReg.pushManager
-      .subscribe({
-        userVisibleOnly: true,
-      })
-      .then((newSub) => {
-        let newSubData = newSub.toJSON()
-        let newSubString = JSON.stringify(newSubData)
-
-        return axiosInstance.post(
-          `${process.env.NEXT_PUBLIC_API_URL}/api/v1/push`,
-          {
-            subscription: newSubString,
-          },
-        )
-      })
-      .then((res) => {
-        console.log(res)
-        displayNotification()
-      })
-
-    // 추후 서버 셋팅 한 뒤 사용
-    // const vapidPublicKey = webpush.generateVAPIDKeys().publicKey
-    // const convertedVapidPublicKey = urlBase64ToUint8Array(vapidPublicKey)
-    // swReg.pushManager
-    //   .subscribe({
-    //     userVisibleOnly: true,
-    //     applicationServerKey: convertedVapidPublicKey,
-    //   })
-    //   .then((newSub) => {
-    //     let newSubData = newSub.toJSON()
-    //     let newSubString = JSON.stringify(newSubData)
-
-    //     return axiosInstance.post(
-    //       `${process.env.NEXT_PUBLIC_API_URL}/api/v1/push`,
-    //       {
-    //         subscription: newSubString,
-    //       },
-    //     )
-    //   })
-    //   .then((res) => {
-    //     console.log(res)
-    //     displayNotification()
-    //   })
-    //   .catch((err) => {
-    //     console.log(err)
-    //   })
-  }
-
-  const handlePushNotification = () => {
-    if ('serviceWorker' in navigator && 'PushManager' in window) {
-      let reg: ServiceWorkerRegistration
-
-      navigator.serviceWorker.ready
-        .then((swReg) => {
-          reg = swReg
-          return swReg.pushManager.getSubscription()
-        })
-        .then((subscription) => {
-          if (subscription === null) {
-            createPushSubscription(reg)
-          }
-        })
-    }
-  }
-
-  const handlePush = () => {
-    if ('PushManager' in window) {
-      Notification.requestPermission((permission) => {
-        if (permission === 'granted') {
-          console.log('Notification permission granted.')
-          handlePushNotification()
-          setIsShowPush(false)
-          localStorage.setItem('isShowPush', 'false')
-        } else {
-          console.log('Unable to get permission to notify.')
-        }
-      })
-    }
-  }
-
   useEffect(() => {
     if (!newData || !newData?.content) return
     //newData가 있어야 업데이트
@@ -289,46 +132,8 @@ const MainPage = ({ initData }: { initData: IPagination<IPost[]> }) => {
 
   return (
     <>
-      {isShowInstall ? (
-        <Box height={'50px'} border="1px solid black">
-          <Stack>
-            사용하시는 브라우저는 PWA 기능을 사용할 수 있습니다.{' '}
-            {isPc ? '데스크탑' : '모바일'}에 설치하시겠습니까?
-            <Stack direction="row">
-              <Button onClick={handleInstall}>설치</Button>
-              <Button onClick={() => setIsShowInstall(false)}>다음에</Button>
-              <Button
-                onClick={() => {
-                  setIsShowInstall(false)
-                  localStorage.setItem('isShowInstall', 'false')
-                }}
-              >
-                닫기
-              </Button>
-            </Stack>
-          </Stack>
-        </Box>
-      ) : null}
-      {isShowPush ? (
-        <Box height={'50px'} border="1px solid black">
-          <Stack>
-            사용하시는 브라우저는 알림 기능을 사용할 수 있습니다.
-            사용하시겠습니까?
-            <Stack direction="row">
-              <Button onClick={handlePush}>네</Button>
-              <Button onClick={() => setIsShowPush(false)}>다음에</Button>
-              <Button
-                onClick={() => {
-                  setIsShowPush(false)
-                  localStorage.setItem('isShowPush', 'false')
-                }}
-              >
-                아니요
-              </Button>
-            </Stack>
-          </Stack>
-        </Box>
-      ) : null}
+      <PushAlertBanner />
+
       {/* mobile view */}
       <Container className="mobile-layout">
         <Box sx={{ backgroundColor: 'white' }} border="1px solid black">
@@ -448,6 +253,7 @@ const MainPage = ({ initData }: { initData: IPagination<IPost[]> }) => {
           </Stack>
         </Stack>
       </Container>
+      <PwaInstallBanner />
     </>
   )
 }
